abstracted request logic into a seperate "request.js" file
